feat: resolve template var in ad tag annotations
