+++ comp/core/autodiscovery/configresolver/configresolver.go
@@ -9,40 +9,19 @@
 package configresolver
 
 import (
-	"encoding/json"
 	"errors"
 	"fmt"
-	"net/url"
-	"os"
-	"regexp"
-	"slices"
 	"sort"
-	"strconv"
-	"strings"
 
 	"github.com/DataDog/datadog-agent/comp/core/autodiscovery/integration"
 	"github.com/DataDog/datadog-agent/comp/core/autodiscovery/listeners"
 	"github.com/DataDog/datadog-agent/comp/core/autodiscovery/providers/names"
 	filter "github.com/DataDog/datadog-agent/comp/core/workloadfilter/def"
-	apiutil "github.com/DataDog/datadog-agent/pkg/api/util"
-	pkgconfigsetup "github.com/DataDog/datadog-agent/pkg/config/setup"
-
 	"github.com/DataDog/datadog-agent/pkg/util/log"
-
-	yaml "gopkg.in/yaml.v2"
+	"github.com/DataDog/datadog-agent/pkg/util/tmplvar"
 )
 
-type variableGetter func(key string, svc listeners.Service) (string, error)
-
-var templateVariables = map[string]variableGetter{
-	"host":     getHost,
-	"pid":      getPid,
-	"port":     getPort,
-	"hostname": getHostname,
-	"env":      getEnvvar,
-	"extra":    getAdditionalTplVariables,
-	"kube":     getAdditionalTplVariables,
-}
+// Template variable resolution is now in pkg/util/tmplvar package
 
 // NoServiceError represents an error that indicates that there's a problem with a service
 type NoServiceError struct {
@@ -124,6 +103,27 @@ func Resolve(tpl integration.Config, svc listeners.Service) (integration.Config,
 	return resolvedConfig, nil
 }
 
+type resolvableService struct {
+	listeners.Service
+}
+
+func (s *resolvableService) GetPorts() ([]tmplvar.ContainerPort, error) {
+	containerPorts, err := s.Service.GetPorts()
+	if err != nil {
+		return nil, err
+	}
+
+	// transform listeners.ContainerPort to tmplvar.ContainerPort
+	tmplContainerPort := make([]tmplvar.ContainerPort, 0, len(containerPorts))
+	for _, contPort := range containerPorts {
+		tmplContainerPort = append(tmplContainerPort, tmplvar.ContainerPort{
+			Name: contPort.Name,
+			Port: contPort.Port,
+		})
+	}
+	return tmplContainerPort, nil
+}
+
 // substituteTemplateVariables replaces %%VARIABLES%% in the config init,
 // instances, and logs config.
 // When there is an error, it stops processing.
@@ -140,7 +140,11 @@ func substituteTemplateVariables(config *integration.Config, svc listeners.Servi
 		if toResolve.dtype == dataInstance {
 			pp = postProcessor
 		}
-		*toResolve.data, err = resolveDataWithTemplateVars(*toResolve.data, svc, toResolve.parser, pp)
+		var resolvableSvc tmplvar.Resolvable
+		if svc != nil {
+			resolvableSvc = &resolvableService{svc}
+		}
+		*toResolve.data, err = tmplvar.ResolveDataWithTemplateVars(*toResolve.data, resolvableSvc, toResolve.parser, pp)
 		if err != nil {
 			return err
 		}
@@ -157,25 +161,10 @@ const (
 	dataLogs
 )
 
-type parser struct {
-	marshal   func(interface{}) ([]byte, error)
-	unmarshal func([]byte, interface{}) error
-}
-
-var jsonp = parser{
-	marshal:   json.Marshal,
-	unmarshal: json.Unmarshal,
-}
-
-var yamlp = parser{
-	marshal:   yaml.Marshal,
-	unmarshal: yaml.Unmarshal,
-}
-
 type dataToResolve struct {
 	data   *integration.Data
 	dtype  dataType
-	parser parser
+	parser tmplvar.Parser
 }
 
 func listDataToResolve(config *integration.Config) []dataToResolve {
@@ -183,7 +172,7 @@ func listDataToResolve(config *integration.Config) []dataToResolve {
 		{
 			data:   &config.InitConfig,
 			dtype:  dataInit,
-			parser: yamlp,
+			parser: tmplvar.YAMLParser,
 		},
 	}
 
@@ -191,14 +180,14 @@ func listDataToResolve(config *integration.Config) []dataToResolve {
 		res = append(res, dataToResolve{
 			data:   &config.Instances[i],
 			dtype:  dataInstance,
-			parser: yamlp,
+			parser: tmplvar.YAMLParser,
 		})
 	}
 
 	if config.IsLogConfig() {
-		p := yamlp
+		p := tmplvar.YAMLParser
 		if config.Provider == names.Container || config.Provider == names.Kubernetes || config.Provider == names.KubeContainer {
-			p = jsonp
+			p = tmplvar.JSONParser
 		}
 		res = append(res, dataToResolve{
 			data:   &config.LogsConfig,
@@ -208,249 +197,6 @@ func listDataToResolve(config *integration.Config) []dataToResolve {
 	}
 
 	return res
-}
-
-func resolveDataWithTemplateVars(data integration.Data, svc listeners.Service, parser parser, postProcessor func(interface{}) error) ([]byte, error) {
-	if len(data) == 0 {
-		return nil, nil
-	}
-
-	var tree interface{}
-
-	// Percent character is not allowed in unquoted yaml strings.
-	data2 := strings.ReplaceAll(string(data), "%%", "‰")
-	if err := parser.unmarshal([]byte(data2), &tree); err != nil {
-		return data, err
-	}
-
-	type treePointer struct {
-		get func() interface{}
-		set func(interface{})
-	}
-
-	stack := []treePointer{
-		{
-			get: func() interface{} {
-				return tree
-			},
-			set: func(x interface{}) {
-				tree = x
-			},
-		},
-	}
-
-	for len(stack) > 0 {
-		n := len(stack) - 1
-		top := stack[n]
-		stack = stack[:n]
-
-		switch elem := top.get().(type) {
-
-		case map[interface{}]interface{}:
-			for k, v := range elem {
-				k2, v2 := k, v
-				stack = append(stack, treePointer{
-					get: func() interface{} {
-						return v2
-					},
-					set: func(x interface{}) {
-						elem[k2] = x
-					},
-				})
-			}
-
-		case map[string]interface{}:
-			for k, v := range elem {
-				k2, v2 := k, v
-				stack = append(stack, treePointer{
-					get: func() interface{} {
-						return v2
-					},
-					set: func(x interface{}) {
-						elem[k2] = x
-					},
-				})
-			}
-
-		case []interface{}:
-			for i, v := range elem {
-				i2, v2 := i, v
-				stack = append(stack, treePointer{
-					get: func() interface{} {
-						return v2
-					},
-					set: func(x interface{}) {
-						elem[i2] = x
-					},
-				})
-			}
-
-		case string:
-			s, err := resolveStringWithTemplateVars(elem, svc)
-			if err != nil {
-				return data, err
-			}
-			// If a `‰` character hasn’t been consumed by a `%%var%%` template variable replacement,
-			// let’s restore it to the initial `%%` value it had in the original string.
-			if str, ok := s.(string); ok {
-				s = strings.ReplaceAll(str, "‰", "%%")
-			}
-			top.set(s)
-
-		case nil, int, bool:
-
-		default:
-			log.Errorf("Unknown type: %T", elem)
-		}
-	}
-
-	if postProcessor != nil {
-		if err := postProcessor(tree); err != nil {
-			return data, err
-		}
-	}
-
-	return parser.marshal(&tree)
-}
-
-// resolveStringWithTemplateVars takes a string as input and replaces all the `‰var_param‰` patterns by the value returned by the appropriate variable getter.
-// It delegates all the work to resolveStringWithAdHocTemplateVars and implements only the following trick:
-// for `‰host‰` patterns, if the value of the variable is an IPv6 *and* it appears in an URL context, then it is surrounded by square brackets.
-// Indeed, IPv6 needs to be surrounded by square brackets inside URL to distinguish the colons of the IPv6 itself from the one separating the IP from the port
-// like in: http://[::1]:80/
-func resolveStringWithTemplateVars(in string, svc listeners.Service) (out interface{}, err error) {
-	isThereAnIPv6Host := false
-
-	adHocTemplateVars := make(map[string]variableGetter)
-	for k, v := range templateVariables {
-		if k == "host" {
-			adHocTemplateVars[k] = func(tplVar string, svc listeners.Service) (string, error) {
-				host, err := getHost(tplVar, svc)
-				if apiutil.IsIPv6(host) {
-					isThereAnIPv6Host = true
-					if tplVar != "" {
-						return fmt.Sprintf("‰host_%s‰", tplVar), nil
-					}
-					return "‰host‰", nil
-				}
-				return host, err
-			}
-		} else {
-			adHocTemplateVars[k] = v
-		}
-	}
-	resolvedString, err := resolveStringWithAdHocTemplateVars(in, svc, adHocTemplateVars)
-	if err != nil {
-		return resolvedString, err
-	}
-
-	if !isThereAnIPv6Host {
-		return resolvedString, err
-	}
-
-	if _, isString := resolvedString.(string); !isString {
-		return resolvedString, err
-	}
-
-	adHocTemplateVars = map[string]variableGetter{
-		"host": func(_ string, _ listeners.Service) (string, error) {
-			return "127.0.0.1", nil
-		},
-	}
-	resolvedStringWithFakeIPv4, err := resolveStringWithAdHocTemplateVars(resolvedString.(string), svc, adHocTemplateVars)
-	if err != nil {
-		return resolvedString, err
-	}
-
-	_, err = url.Parse(resolvedStringWithFakeIPv4.(string))
-	if err != nil {
-		return resolvedString, nil
-	}
-
-	adHocTemplateVars = map[string]variableGetter{
-		"host": func(tplVar string, svc listeners.Service) (string, error) {
-			host, err := getHost(tplVar, svc)
-			var sb strings.Builder
-			sb.WriteByte('[')
-			sb.WriteString(host)
-			sb.WriteByte(']')
-			return sb.String(), err
-		},
-	}
-	resolvedStringWithIPv6, err := resolveStringWithAdHocTemplateVars(resolvedString.(string), svc, adHocTemplateVars)
-	if err != nil {
-		return resolvedString, err
-	}
-
-	_, err = url.Parse(resolvedStringWithIPv6.(string))
-	if err != nil {
-		return resolveStringWithAdHocTemplateVars(in, svc, templateVariables)
-	}
-
-	return resolvedStringWithIPv6, err
-}
-
-var varPattern = regexp.MustCompile(`‰(.+?)(?:_(.+?))?‰`)
-
-// resolveStringWithAdHocTemplateVars takes a string as input and replaces all the `‰var_param‰` patterns by the value returned by the appropriate variable getter.
-// The variable getters are passed as last parameter.
-// If the input string is composed of *only* a `‰var_param‰` pattern and the result of the substitution is a boolean or a number, then the function returns a boolean or a number instead of a string.
-func resolveStringWithAdHocTemplateVars(in string, svc listeners.Service, templateVariables map[string]variableGetter) (out interface{}, err error) {
-	varIndexes := varPattern.FindAllStringSubmatchIndex(in, -1)
-
-	if len(varIndexes) == 0 {
-		return in, nil
-	}
-
-	var sb strings.Builder
-
-	sb.WriteString(in[:varIndexes[0][0]])
-	for i := range varIndexes {
-		if i != 0 {
-			sb.WriteString(in[varIndexes[i-1][1]:varIndexes[i][0]])
-		}
-
-		varName := in[varIndexes[i][2]:varIndexes[i][3]]
-		varKey := ""
-		if varIndexes[i][4] != -1 {
-			varKey = in[varIndexes[i][4]:varIndexes[i][5]]
-		}
-
-		if f, found := templateVariables[varName]; found {
-			resolvedVar, e := f(varKey, svc)
-			if e != nil {
-				err = e
-			}
-			sb.WriteString(resolvedVar)
-		} else {
-			endTagIdx := varIndexes[i][5]
-			if endTagIdx == -1 {
-				endTagIdx = varIndexes[i][3]
-			}
-			err := fmt.Errorf("invalid %%%%%s%%%% tag", in[varIndexes[i][2]:endTagIdx])
-			if svc != nil {
-				err = fmt.Errorf("unable to add tags for service '%s', err: %w", svc.GetServiceID(), err)
-			}
-			return out, err
-		}
-	}
-	sb.WriteString(in[varIndexes[len(varIndexes)-1][1]:])
-
-	out = sb.String()
-
-	if len(varIndexes) == 1 &&
-		varIndexes[0][0] == 0 &&
-		varIndexes[0][1] == len(in) {
-
-		if i, e := strconv.ParseInt(out.(string), 0, 64); e == nil {
-			return i, err
-		}
-		if b, e := strconv.ParseBool(out.(string)); e == nil {
-			return b, err
-		}
-	}
-
-	return
 }
 
 func tagsAdder(tags []string) func(interface{}) error {
@@ -496,171 +242,5 @@ func tagsAdder(tags []string) func(interface{}) error {
 	}
 }
 
-func getHost(tplVar string, svc listeners.Service) (string, error) {
-	if svc == nil {
-		return "", newNoServiceError("No service. %%%%host%%%% is not allowed")
-	}
-
-	hosts, err := svc.GetHosts()
-	if err != nil {
-		return "", fmt.Errorf("failed to extract IP address for container %s, ignoring it. Source error: %s", svc.GetServiceID(), err)
-	}
-	if len(hosts) == 0 {
-		return "", fmt.Errorf("no network found for container %s, ignoring it", svc.GetServiceID())
-	}
-
-	// a network was specified
-	if ip, ok := hosts[tplVar]; ok {
-		return ip, nil
-	}
-	log.Debugf("Network %q not found, trying bridge IP instead", tplVar)
-
-	// otherwise use fallback policy
-	ip, err := getFallbackHost(hosts)
-	if err != nil {
-		return "", fmt.Errorf("failed to resolve IP address for container %s, ignoring it. Source error: %s", svc.GetServiceID(), err)
-	}
-
-	return ip, nil
-}
-
-// getFallbackHost implements the fallback strategy to get a service's IP address
-// the current strategy is:
-//   - if there's only one network we use its IP
-//   - otherwise we look for the bridge net and return its IP address
-//   - if we can't find it we fail because we shouldn't try and guess the IP address
-func getFallbackHost(hosts map[string]string) (string, error) {
-	if len(hosts) == 1 {
-		for _, host := range hosts {
-			return host, nil
-		}
-	}
-
-	bridgeIP, bridgeIsPresent := hosts["bridge"]
-	if bridgeIsPresent {
-		return bridgeIP, nil
-	}
-
-	return "", errors.New("not able to determine which network is reachable")
-}
-
-// getPort returns ports of the service
-func getPort(tplVar string, svc listeners.Service) (string, error) {
-	if svc == nil {
-		return "", newNoServiceError("No service. %%%%port%%%% is not allowed")
-	}
-
-	ports, err := svc.GetPorts()
-	if err != nil {
-		return "", fmt.Errorf("failed to extract port list for container %s, ignoring it. Source error: %s", svc.GetServiceID(), err)
-	} else if len(ports) == 0 {
-		return "", fmt.Errorf("no port found for container %s - ignoring it", svc.GetServiceID())
-	}
-
-	if len(tplVar) == 0 {
-		return strconv.Itoa(ports[len(ports)-1].Port), nil
-	}
-
-	idx, err := strconv.Atoi(tplVar)
-	if err != nil {
-		// The template variable is not an index so try to lookup port by name.
-		for _, port := range ports {
-			if port.Name == tplVar {
-				return strconv.Itoa(port.Port), nil
-			}
-		}
-		return "", fmt.Errorf("port %s not found, skipping container %s", tplVar, svc.GetServiceID())
-	}
-	if len(ports) <= idx {
-		return "", fmt.Errorf("index given for the port template var is too big, skipping container %s", svc.GetServiceID())
-	}
-	return strconv.Itoa(ports[idx].Port), nil
-}
-
-// getPid returns the process identifier of the service
-func getPid(_ string, svc listeners.Service) (string, error) {
-	if svc == nil {
-		return "", newNoServiceError("No service. %%%%pid%%%% is not allowed")
-	}
-
-	pid, err := svc.GetPid()
-	if err != nil {
-		return "", fmt.Errorf("failed to get pid for service %s, skipping config - %s", svc.GetServiceID(), err)
-	}
-	return strconv.Itoa(pid), nil
-}
-
-// getHostname returns the hostname of the service, to be used
-// when the IP is unavailable or erroneous
-func getHostname(_ string, svc listeners.Service) (string, error) {
-	if svc == nil {
-		return "", newNoServiceError("No service. %%%%hostname%%%% is not allowed")
-	}
-
-	name, err := svc.GetHostname()
-	if err != nil {
-		return "", fmt.Errorf("failed to get hostname for service %s, skipping config - %s", svc.GetServiceID(), err)
-	}
-	return name, nil
-}
-
-// getAdditionalTplVariables returns listener-specific template variables.
-// It resolves template variables prefixed with kube_ or extra_
-// Even though it gets the data from the same listener method GetExtraConfig, the kube_ and extra_
-// prefixes are customer facing, we support both of them for a better user experience depending on
-// the AD listener and what the template variable represents.
-func getAdditionalTplVariables(tplVar string, svc listeners.Service) (string, error) {
-	if svc == nil {
-		return "", newNoServiceError("No service. %%%%extra_*%%%% or %%%%kube_*%%%% are not allowed")
-	}
-
-	value, err := svc.GetExtraConfig(tplVar)
-	if err != nil {
-		return "", fmt.Errorf("failed to get extra info for service %s, skipping config - %s", svc.GetServiceID(), err)
-	}
-	return value, nil
-}
-
-// getEnvvar returns a system environment variable if found
-func getEnvvar(envVar string, svc listeners.Service) (string, error) {
-	if len(envVar) == 0 {
-		if svc != nil {
-			return "", fmt.Errorf("envvar name is missing, skipping service %s", svc.GetServiceID())
-		}
-		return "", errors.New("envvar name is missing")
-	}
-
-	if !allowEnvVar(envVar) {
-		if svc != nil {
-			return "", fmt.Errorf("envvar %s is not allowed in check configs, skipping service %s", envVar, svc.GetServiceID())
-		}
-		return "", fmt.Errorf("envvar %s is not allowed in check configs", envVar)
-	}
-
-	value, found := os.LookupEnv(envVar)
-	if !found {
-		if svc != nil {
-			return "", fmt.Errorf("failed to retrieve envvar %s, skipping service %s", envVar, svc.GetServiceID())
-		}
-		return "", fmt.Errorf("failed to retrieve envvar %s", envVar)
-	}
-	return value, nil
-}
-
-func allowEnvVar(envVar string) bool {
-	if pkgconfigsetup.Datadog().GetBool("ad_disable_env_var_resolution") {
-		return false
-	}
-
-	allowedEnvs := pkgconfigsetup.Datadog().GetStringSlice("ad_allowed_env_vars")
-
-	// If the option is not set or is empty, the default behavior applies: all
-	// envs are allowed.
-	if len(allowedEnvs) == 0 {
-		return true
-	}
-
-	return slices.ContainsFunc(allowedEnvs, func(env string) bool {
-		return strings.EqualFold(env, envVar)
-	})
-}
+// All template variable getter functions (GetHost, GetPort, GetPid, GetHostname,
+// GetEnvvar, GetAdditionalTplVariables) have been moved to pkg/util/tmplvar package
